+++ src/archivist/__init__.py
@@ -17,5 +17,5 @@ __init__.py
 """
 
 __app_name__ = "archivist"
-__version__ = "0.1.3"
+__version__ = "0.1.4"
 DEBUG = True