+++ CHANGELOG.md
@@ -7,17 +7,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.4] - 2023-09-23
+
+### Added
+
+- Crawler clones repo to target directory
+
 ## [0.1.3] - 2023-09-23
 
 ### Fixed
 
-- Fixed a bug where the `--version` and `--update` flags were not working as expected, switched to argparse
+- The `--version` and `--update` flags now work as expected, switched to argparse
 
 ## [0.1.2] - 2023-09-23
 
 ### Fixed
 
-- Fixed a bug where the `--version` and `--update` flags were not working as expected
+- The `--version` and `--update` flags should now work as expected
 
 ## [0.1.1] - 2023-09-23